+++ src/encoding.h
@@ -4,7 +4,7 @@
 extern Instruction emptyInstruction;
 
 // Use somente para valores sem sinal (registradores rs, rt e rd, etc)
-uint8_t binaryToUnsignedInt(const char *serializedBinary);
+int binaryToUnsignedInt(const char *serializedBinary);
 
 void extendSignal(char *baseSerializedBinary, char *buffer, int size);
